Switch to using react-router loaders
This is close to simulating SSR with remix, and better separates
concerns for an arbitrary SSR framework in general.

We now do a lot of processing of the zoekt search API response in the
zod transformers, such as parsing `Content` into lines, in order to make
the response more suitable for rendering. This is how we will do things
like syntax highlighting in the future, without the relevant libraries
ever touching the client. Zod itself will not be on the client when we
start doing SSR.

This change removes live search debouncing. It mostly didn't work, and
the True Way here is to ensure that API responses are small enough to be
downloaded and rendered for any search. This is not the case with small
search queries today, but there's no way we can fix that with zoekt
parameters. We will need to truncate zoekt responses in the loader,
which is a big bag of worms.

Fix a few minor bugs:
- Inline some effects into callbacks.  In the case of
  NonNegativeIntegerInput, this fixed a newly revealed render loop.
- Generalize that component to IntegerInput, as it was a bug to allow
  files=0, which zoekt interprets as no limit.  files must be positive.
- Continue to show the lander on search error if there are no previous
  results.