Reduce default `files` from 50 to 20
We need to improve performance as the first few characters are typed. I
have upcoming changes to reduce the number of matches in the DOM in
situations where there are very large chunks (common with just 1 or 2
character searches), which helps a little bit.

But the UI lag when typing the first few characters is ultimately driven
by JS time, parsing the JSON results and processing the huge number of
matches per file produced by such queries. Barring an API for capping
the number of _matches_, reducing the number of files is the only viable
option. (Such an API is possible, but only if we are running something
custom in zoekt.)

50 as a default was always kind of overkill anyway, purely from the
number of results displayed. It can be trivially dialed up when users
need more.