Improve "cutoff" logic in the worst cases, add automated tests
In certain cases, especially in the first few characters of a query, you
can get many thousands of matches in a file. And large numbers of these
matches can be within a single chunk. We have to cut these chunks off to
not end up rendering dozens or hundreds of lines of a file when that
file is collapsed.

I'm doing this with a somewhat involved soft/hard cutoff; there are
other strategies to consider, but none of them yield results that are
substantially different. The goal of rendering way fewer matches in the
worst case when files are collapsed is achieved.

Also, finally add support for automated tests to the project. We do a
little snapshotting.