Move search result parsing back to the server
Now that we response sizes from zoekt that are one or two orders of
magnitude smaller than before in the worst case, we can return to
parsing those responses on the server, which reduces the code size and
CPU time spent in the browser.

This will become substantially more consequential when we add syntax
highlighting.