Add syntax highlighting
With [@wooorm/starry-night](https://github.com/wooorm/starry-night).

We have achieved the goal of doing all highlighting on the server, such
that this big/expensive lib does not have to be sent to the browser.

This entails a complete rewrite of content-parser to be string-oriented
instead of bytes-oriented, as starry-night of course operates on
strings. We join starry-night's highlighting data to zoekt's match data
to produce the same kind of line-oriented content we always have.

Overall, starry-night seems to typically not be _that_ expensive, with
the default limits on search results. To highlight a 20 files/500
matches result that I benchmarked took in the low tens of ms over dozens
of invocations (one for each ChunkMatch). Not great, not terrible. The
biggest risk is in what this does to a contended NodeJS server.

It's worth noting that our highlighting strategy is kept optimistic for
simplicity: given that we are only highlighting individual ChunkMatches,
the fidelity of the highlighting is not as high as it could be. The
highlighter can only do so much with limited context on the content
being highlighted; it would surely be able to do better with the full
files. And we could give it the full files, as we can ask zoekt for
them, but then perf would tank, both in terms of highlighting (a 10kloc
C file from the Linux source took over 1s to highlight) and in terms of
wire transfer. And so we'd need some kind of asynchronous highlighting
strategy, probably with caching. I am not interested in adding that kind
of complexity to the project right now.

Even with the optimistic strategy, there is potential for better
highlighting performance in the future, with lazy highlighting of hidden
matches, though that itself would require a lot more complexity in the
UI layer, and might not be worth it from the user's perspective, due to
round trip time.