Clean up file/match count UI
Call a spade a spade; call a zoekt a zoekt; call a neogrok a neogrok.

Remove the false promise of "frontend timings", as it's really tied to
svelte render in ways that make it nonsense (e.g. the number shoots
toward zero on hydration).

Highlight neogrok file/match counts when they are limiting the number of
results.