Improve demo site deployment actions
I really don't want the demo site to be deployed if CI fails, so we need
a dependency between the relevant jobs. Apparently different workflows
are supposed to be totally orthogonal, and creating dependencies between
them is kind of janky.

So, move the deployment jobs into the CI workflow, and make them
appropriately conditional. And use a matrix; why not?