Expand `devalue` bypass to /api/list
Turns out that when you have a lot of repositories, the bad `devalue`
performance starts to bite on the repositories list page just like it
did on the search page.

So, genericize the whole concept and apply it to the new `/api/list`
endpoint just like we did for `/api/search`.