Overhaul /repositories
Increase performance on instances with large number of repositories by
using the same limiting semantics we use on the search page. The
difference is that this limiting is done entirely in neogrok, as zoekt
has no scoring/truncating parameters for repository search.  So this
limit does nothing to reduce API bandwith, only to help render
performance. Rendering an HTML table with 5000 rows is simply too
expensive to do on every keypress.

Make the table sortable by columns.

Clarify the table layout and expand the documentation to describe what
shards are.