Handle breaking change in ChunkMatch newlines
Previously, zoekt made the curious choice to not include trailing
newline characters in each chunk it served in its API, in spite of the
posix definition of a line including the trailing newline character,
which is respected in many line-oriented unix tools like grep/ripgrep,
etc.

This mistake has finally been corrected! But it is a breakage that we
have to handle here. I've updated the tests to reflect the kind of data
that zoekt is actually serving now.

See https://github.com/sourcegraph/zoekt/pull/747.