Set `content-type` on /api/list requests
Oops. The lack of a header resulted in a default of `text/plain` which
spuriously triggers SvelteKit's CSRF prevention machinery.

This bug was around forever but somehow you did not always need to work
around it by setting ORIGIN - I think it could only have been triggered
when visiting neogrok on `localhost`?

Ref #30.