Handle missing optional Repository properties
If zoekt ever starts `omitempty`ing a lot of these optional properties,
neogrok's current validation will reject them. Handle any such cases,
and use the resulting type information to improve the display of such
missing data on /repositories.

The zero-valued `lastCommit` and `branches` bits are actually a problem
now, in that the table cells are just empty for these. I am opting to
instead say "n/a" as I think it makes it clearer what's going on.