Add dark mode UI
This uses the OS appearance `prefers-color-scheme` CSS support in
tailwind. There is no support for an explicit preference right now, and
I'm not sure I'll ever add one.

This preference also drives a svelte store that we use to select the
syntax highlighting theme, such that we use the GitHub dark theme when
dark mode is enabled.

This also involves changing the color scheme for the SVG search
preference buttons, as it's either not easy or not possible to use media
queries to drive SVG fill/stroke colors, and the existing ones only
worked for the light theme. I've changed the color styling of these
buttons to be driven purely by CSS, which has resulted in a materially
different appearance.