Bump dependencies
This is mainly to fix vulnerabilities in SvelteKit and vite, but I went
ahead and did all the tooling.

We go to Svelte 5 but do not use runes yet. Or maybe ever, I'm not sure
I'll put in the effort to migrate anything considering that the project
is "done".

The required nodejs version does not change (we use node 20).